+++ scripts/data.js
@@ -1,18 +1,24 @@
+const firebase = require("firebase/app");
+require("firebase/database");
+import { displayData } from "./dom";
+
 let clients = [];
 
-const db = firebase.database();
-const clientsRef = db.ref("clients");
+export function getData() {
+  console.log("getData");
+  const db = firebase.database();
+  const clientsRef = db.ref("clients");
+  clientsRef.on("value", (snapshot) => {
+    // console.log(snapshot.val());
+    // clients = Array.from(Object.values(snapshot.val()));
+    console.log("Display data will be called");
+    clients = convertObjToArray(snapshot.val());
+    // console.log(Object.keys(clients));
+    // console.log(Object.values(clients));
 
-clientsRef.on("value", (snapshot) => {
-  // console.log(snapshot.val());
-  // clients = Array.from(Object.values(snapshot.val()));
-  console.log("Display data will be called");
-  clients = convertObjToArray(snapshot.val());
-  // console.log(Object.keys(clients));
-  // console.log(Object.values(clients));
-
-  displayData(clients);
-});
+    displayData(clients);
+  });
+}
 
 function convertObjToArray(object) {
   return Object.keys(object).map((key) => {
@@ -36,515 +42,4 @@ function convertObjToArray(object) {
   // });
 }
 
-// const clients = [
-//   {
-//     id: 1,
-//     firstName: "Demetris",
-//     lastName: "Nerheny",
-//     email: "dnerheny0@timesonline.co.uk",
-//     gender: "Male",
-//     amount: "$2.08",
-//     date: "7/28/2019",
-//     avatar: "https://robohash.org/omnisveniamqui.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 2,
-//     firstName: "Ameline",
-//     lastName: "Labusquiere",
-//     email: "alabusquiere1@meetup.com",
-//     gender: "Female",
-//     amount: "$9.85",
-//     date: "5/3/2019",
-//     avatar: "https://robohash.org/sapientedoloret.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 3,
-//     firstName: "Glenn",
-//     lastName: "Eccersley",
-//     email: "geccersley2@google.com",
-//     gender: "Female",
-//     amount: "$5.57",
-//     date: "11/18/2019",
-//     avatar: "https://robohash.org/etnonquam.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 4,
-//     firstName: "Mab",
-//     lastName: "Witchell",
-//     email: "mwitchell3@newsvine.com",
-//     gender: "Female",
-//     amount: "$2.37",
-//     date: "2/10/2019",
-//     avatar: "https://robohash.org/autemutquos.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 5,
-//     firstName: "Otha",
-//     lastName: "Parlet",
-//     email: "oparlet4@angelfire.com",
-//     gender: "Female",
-//     amount: "$8.99",
-//     date: "3/14/2019",
-//     avatar: "https://robohash.org/beataecorruptiut.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 6,
-//     firstName: "Udale",
-//     lastName: "Winslet",
-//     email: "uwinslet5@facebook.com",
-//     gender: "Male",
-//     amount: "$1.91",
-//     date: "10/23/2019",
-//     avatar: "https://robohash.org/nesciuntiustoqui.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 7,
-//     firstName: "Chrissie",
-//     lastName: "Brecher",
-//     email: "cbrecher6@nasa.gov",
-//     gender: "Female",
-//     amount: "$9.02",
-//     date: "10/7/2019",
-//     avatar: "https://robohash.org/etdelenitiqui.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 8,
-//     firstName: "Whitaker",
-//     lastName: "Stalman",
-//     email: "wstalman7@bloglines.com",
-//     gender: "Male",
-//     amount: "$6.05",
-//     date: "3/13/2019",
-//     avatar:
-//       "https://robohash.org/voluptatemrerumpraesentium.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 9,
-//     firstName: "Anthea",
-//     lastName: "Shapcott",
-//     email: "ashapcott8@barnesandnoble.com",
-//     gender: "Female",
-//     amount: "$7.41",
-//     date: "6/28/2019",
-//     avatar: "https://robohash.org/etetlaudantium.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 10,
-//     firstName: "Risa",
-//     lastName: "Bedder",
-//     email: "rbedder9@myspace.com",
-//     gender: "Female",
-//     amount: "$2.46",
-//     date: "3/15/2019",
-//     avatar: "https://robohash.org/repudiandaeeumrerum.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 11,
-//     firstName: "Granville",
-//     lastName: "Antonio",
-//     email: "gantonioa@aboutads.info",
-//     gender: "Male",
-//     amount: "$2.07",
-//     date: "8/7/2019",
-//     avatar: "https://robohash.org/autemdolorenesciunt.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 12,
-//     firstName: "Walton",
-//     lastName: "Dexter",
-//     email: "wdexterb@ehow.com",
-//     gender: "Male",
-//     amount: "$8.58",
-//     date: "7/1/2019",
-//     avatar: "https://robohash.org/modivoluptateunde.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 13,
-//     firstName: "Thaddus",
-//     lastName: "Lisett",
-//     email: "tlisettc@reference.com",
-//     gender: "Male",
-//     amount: "$0.99",
-//     date: "10/27/2019",
-//     avatar: "https://robohash.org/aliasquiet.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 14,
-//     firstName: "Lorita",
-//     lastName: "Mc Ilory",
-//     email: "lmciloryd@chron.com",
-//     gender: "Female",
-//     amount: "$3.16",
-//     date: "3/20/2019",
-//     avatar: "https://robohash.org/quosnonsimilique.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 15,
-//     firstName: "Gareth",
-//     lastName: "Esch",
-//     email: "gesche@liveinternet.ru",
-//     gender: "Male",
-//     amount: "$5.53",
-//     date: "11/7/2019",
-//     avatar:
-//       "https://robohash.org/officiiscorporissuscipit.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 16,
-//     firstName: "Pail",
-//     lastName: "Rohfsen",
-//     email: "prohfsenf@phoca.cz",
-//     gender: "Male",
-//     amount: "$9.50",
-//     date: "1/6/2019",
-//     avatar: "https://robohash.org/estodiout.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 17,
-//     firstName: "Linda",
-//     lastName: "Villaron",
-//     email: "lvillarong@tuttocitta.it",
-//     gender: "Female",
-//     amount: "$9.93",
-//     date: "12/19/2018",
-//     avatar:
-//       "https://robohash.org/architectoeumveritatis.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 18,
-//     firstName: "Alisander",
-//     lastName: "Abbatucci",
-//     email: "aabbatuccih@mozilla.com",
-//     gender: "Male",
-//     amount: "$7.61",
-//     date: "9/24/2019",
-//     avatar: "https://robohash.org/debitisexcepturiquia.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 19,
-//     firstName: "Belva",
-//     lastName: "Mingardo",
-//     email: "bmingardoi@techcrunch.com",
-//     gender: "Female",
-//     amount: "$3.40",
-//     date: "7/10/2019",
-//     avatar: "https://robohash.org/oditvelitnon.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 20,
-//     firstName: "Candida",
-//     lastName: "Drummer",
-//     email: "cdrummerj@un.org",
-//     gender: "Female",
-//     amount: "$6.32",
-//     date: "4/11/2019",
-//     avatar: "https://robohash.org/explicaboquosvel.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 21,
-//     firstName: "Leslie",
-//     lastName: "Jakab",
-//     email: "ljakabk@seesaa.net",
-//     gender: "Male",
-//     amount: "$7.29",
-//     date: "5/17/2019",
-//     avatar: "https://robohash.org/hiceiusmollitia.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 22,
-//     firstName: "Gussi",
-//     lastName: "Bratley",
-//     email: "gbratleyl@adobe.com",
-//     gender: "Female",
-//     amount: "$2.91",
-//     date: "12/4/2018",
-//     avatar: "https://robohash.org/accusantiumatfugiat.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 23,
-//     firstName: "Nicky",
-//     lastName: "Hollow",
-//     email: "nhollowm@nifty.com",
-//     gender: "Male",
-//     amount: "$4.04",
-//     date: "7/10/2019",
-//     avatar: "https://robohash.org/reiciendisesseea.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 24,
-//     firstName: "Evita",
-//     lastName: "Raysdale",
-//     email: "eraysdalen@diigo.com",
-//     gender: "Female",
-//     amount: "$2.91",
-//     date: "10/28/2019",
-//     avatar:
-//       "https://robohash.org/doloremqueassumendaexcepturi.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 25,
-//     firstName: "Saidee",
-//     lastName: "Seathwright",
-//     email: "sseathwrighto@i2i.jp",
-//     gender: "Female",
-//     amount: "$3.18",
-//     date: "6/4/2019",
-//     avatar: "https://robohash.org/quasquiculpa.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 26,
-//     firstName: "Joni",
-//     lastName: "Paty",
-//     email: "jpatyp@so-net.ne.jp",
-//     gender: "Female",
-//     amount: "$4.68",
-//     date: "4/3/2019",
-//     avatar: "https://robohash.org/nonvoluptatibusvel.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 27,
-//     firstName: "Marc",
-//     lastName: "Laister",
-//     email: "mlaisterq@washington.edu",
-//     gender: "Male",
-//     amount: "$2.57",
-//     date: "6/18/2019",
-//     avatar: "https://robohash.org/rerummodicorrupti.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 28,
-//     firstName: "Casandra",
-//     lastName: "Douty",
-//     email: "cdoutyr@linkedin.com",
-//     gender: "Female",
-//     amount: "$3.41",
-//     date: "9/24/2019",
-//     avatar: "https://robohash.org/necessitatibusfugaet.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 29,
-//     firstName: "Ramsey",
-//     lastName: "Grisewood",
-//     email: "rgrisewoods@addtoany.com",
-//     gender: "Male",
-//     amount: "$4.58",
-//     date: "11/9/2019",
-//     avatar: "https://robohash.org/autquosrecusandae.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 30,
-//     firstName: "Evyn",
-//     lastName: "Rehor",
-//     email: "erehort@rambler.ru",
-//     gender: "Male",
-//     amount: "$2.43",
-//     date: "7/10/2019",
-//     avatar: "https://robohash.org/quassitvoluptate.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 31,
-//     firstName: "Clarette",
-//     lastName: "Bellenie",
-//     email: "cbellenieu@imgur.com",
-//     gender: "Female",
-//     amount: "$5.22",
-//     date: "5/27/2019",
-//     avatar:
-//       "https://robohash.org/temporaassumendapossimus.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 32,
-//     firstName: "Christiana",
-//     lastName: "Olpin",
-//     email: "colpinv@163.com",
-//     gender: "Female",
-//     amount: "$9.59",
-//     date: "12/22/2018",
-//     avatar: "https://robohash.org/uttemporehic.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 33,
-//     firstName: "Brande",
-//     lastName: "Ovize",
-//     email: "bovizew@theguardian.com",
-//     gender: "Female",
-//     amount: "$3.10",
-//     date: "11/10/2019",
-//     avatar: "https://robohash.org/nullaoditdeleniti.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 34,
-//     firstName: "Garret",
-//     lastName: "Marriot",
-//     email: "gmarriotx@meetup.com",
-//     gender: "Male",
-//     amount: "$6.67",
-//     date: "7/10/2019",
-//     avatar: "https://robohash.org/iddelenitiexcepturi.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 35,
-//     firstName: "Amory",
-//     lastName: "Janssen",
-//     email: "ajansseny@so-net.ne.jp",
-//     gender: "Male",
-//     amount: "$5.46",
-//     date: "4/18/2019",
-//     avatar: "https://robohash.org/minusnobisquasi.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 36,
-//     firstName: "Sholom",
-//     lastName: "Stroton",
-//     email: "sstrotonz@about.me",
-//     gender: "Male",
-//     amount: "$4.67",
-//     date: "8/3/2019",
-//     avatar: "https://robohash.org/sitnisiet.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 37,
-//     firstName: "Angeline",
-//     lastName: "Tevelov",
-//     email: "atevelov10@edublogs.org",
-//     gender: "Female",
-//     amount: "$3.57",
-//     date: "9/25/2019",
-//     avatar:
-//       "https://robohash.org/excepturivoluptatemesse.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 38,
-//     firstName: "Eleanor",
-//     lastName: "Micallef",
-//     email: "emicallef11@hugedomains.com",
-//     gender: "Female",
-//     amount: "$4.47",
-//     date: "6/3/2019",
-//     avatar: "https://robohash.org/rerumharumet.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 39,
-//     firstName: "Ilene",
-//     lastName: "Kellen",
-//     email: "ikellen12@dyndns.org",
-//     gender: "Female",
-//     amount: "$4.54",
-//     date: "9/17/2019",
-//     avatar: "https://robohash.org/sitoditrerum.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 40,
-//     firstName: "Chalmers",
-//     lastName: "Pearton",
-//     email: "cpearton13@sphinn.com",
-//     gender: "Male",
-//     amount: "$9.46",
-//     date: "10/12/2019",
-//     avatar:
-//       "https://robohash.org/explicabovoluptatemodi.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 41,
-//     firstName: "Caz",
-//     lastName: "Astley",
-//     email: "castley14@blogspot.com",
-//     gender: "Male",
-//     amount: "$3.76",
-//     date: "8/6/2019",
-//     avatar:
-//       "https://robohash.org/aspernaturetnecessitatibus.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 42,
-//     firstName: "Gram",
-//     lastName: "Culverhouse",
-//     email: "gculverhouse15@home.pl",
-//     gender: "Male",
-//     amount: "$3.49",
-//     date: "2/26/2019",
-//     avatar:
-//       "https://robohash.org/consequaturinciduntconsequatur.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 43,
-//     firstName: "Drona",
-//     lastName: "Muckeen",
-//     email: "dmuckeen16@cisco.com",
-//     gender: "Female",
-//     amount: "$0.31",
-//     date: "7/27/2019",
-//     avatar: "https://robohash.org/minusestquos.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 44,
-//     firstName: "Haleigh",
-//     lastName: "Stendell",
-//     email: "hstendell17@china.com.cn",
-//     gender: "Male",
-//     amount: "$4.87",
-//     date: "12/12/2018",
-//     avatar: "https://robohash.org/fugiatipsaomnis.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 45,
-//     firstName: "Jade",
-//     lastName: "Miettinen",
-//     email: "jmiettinen18@jigsy.com",
-//     gender: "Female",
-//     amount: "$3.27",
-//     date: "8/4/2019",
-//     avatar:
-//       "https://robohash.org/doloremrationeeligendi.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 46,
-//     firstName: "Grange",
-//     lastName: "Rheaume",
-//     email: "grheaume19@rambler.ru",
-//     gender: "Male",
-//     amount: "$9.61",
-//     date: "10/17/2019",
-//     avatar: "https://robohash.org/sedsuntet.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 47,
-//     firstName: "Merna",
-//     lastName: "Peetermann",
-//     email: "mpeetermann1a@unc.edu",
-//     gender: "Female",
-//     amount: "$6.54",
-//     date: "4/25/2019",
-//     avatar: "https://robohash.org/indistinctiodicta.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 48,
-//     firstName: "Marice",
-//     lastName: "Full",
-//     email: "mfull1b@hao123.com",
-//     gender: "Female",
-//     amount: "$1.07",
-//     date: "11/15/2019",
-//     avatar: "https://robohash.org/quivitaeipsum.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 49,
-//     firstName: "Neils",
-//     lastName: "Gourley",
-//     email: "ngourley1c@examiner.com",
-//     gender: "Male",
-//     amount: "$3.70",
-//     date: "5/29/2019",
-//     avatar: "https://robohash.org/doloreliberofacere.jpg?size=50x50&set=set1"
-//   },
-//   {
-//     id: 50,
-//     firstName: "Issie",
-//     lastName: "Skill",
-//     email: "iskill1d@linkedin.com",
-//     gender: "Female",
-//     amount: "$9.18",
-//     date: "9/7/2019",
-//     avatar: "https://robohash.org/autnatuseum.jpg?size=50x50&set=set1"
-//   }
-// ];
+export { clients };
