+++ scripts/script.js
@@ -1,19 +1,26 @@
+import "bootstrap";
+import "bootstrap/dist/css/bootstrap.min.css";
+import "../styles/styles.css";
+
+window.jQuery = window.$ = require("jquery");
+
+const firebase = require("firebase/app");
+require("firebase/auth");
+require("firebase/database");
+import { initApp } from "./firebase.js";
+import { getData, clients } from "./data";
+import { refreshData, showResultListOrNotFound } from "./dom";
+
+initApp();
+getData();
+
 //Observe changes
 firebase.auth().onAuthStateChanged((user) => {
   if (user) {
-    // User is signed in.
-    // let displayName = user.displayName;
     let email = user.email;
-    // alert("Hello " + email);
-    // let emailVerified = user.emailVerified;
-    // let photoURL = user.photoURL;
-    // let isAnonymous = user.isAnonymous;
-    // let uid = user.uid;
-    // let providerData = user.providerData;
-    // ...
   } else {
     // User is signed out.
-    window.location.href = "login.html";
+    window.location.href = "./login.html";
   }
 });
 
@@ -35,108 +42,25 @@ const editClientForm = document.querySelector("#editClientForm");
 editClientForm.addEventListener("submit", (event) => {
   event.preventDefault();
   // console.log(event.target);
-  editClient(event.target); //TODO: add Edit client form
+  editClient(event.target); //TODO: comment
   hideModel("#editClientForm");
 });
 
-function displayData(clientsList = clients) {
-  clearList();
-  const ul = document.querySelector("#clientsData");
-
-  clientsList.forEach((client) => {
-    ul.appendChild(getLiElement(client));
+const sortsFields = [
+  { id: "sortAscending", value: "ascending" },
+  { id: "sortDescending", value: "descending" },
+];
+sortsFields.forEach((field) => {
+  const element = document.querySelector(`#${field.id}`);
+  element.addEventListener("click", () => {
+    sortList(field.value);
   });
-  // for (const property in clientsList) {
-  //   // console.log(`${property}: ${clientsList[property]}`);
-  //   ul.appendChild(getLiElement(clientsList[property], property));
-  // }
+});
 
-  // sumAmount(Array.from(Object.values(clientsList)));
-  sumAmount(clientsList);
-}
-
-function getLiElement(client) {
-  const { avatar, clientId } = client;
-  const newLi = document.createElement("li");
-  const image = document.createElement("img");
-  newLi.className = "media";
-  newLi.id = clientId;
-  image.className = "mr-3 align-self-center";
-  image.setAttribute("src", avatar);
-
-  newLi.appendChild(image);
-  newLi.appendChild(createClientDescription(client, clientId));
-  return newLi;
-}
-
-function createClientDescription(client, id) {
-  const div = document.createElement("div");
-  div.className = "media-body";
-
-  const mailLink = document.createElement("a");
-  mailLink.setAttribute("href", `mailto:${client.email}`);
-  mailLink.innerHTML = client.email;
-
-  const textPart1 = document.createTextNode(
-    `${client.lastName} ${client.firstName} - `
-  );
-
-  const textPart2 = document.createTextNode(
-    ` ${client.gender} (${client.date} - ${client.amount})`
-  );
-
-  const deleteLink = document.createElement("a");
-  deleteLink.innerHTML = "Delete";
-  deleteLink.setAttribute("href", "#");
-  deleteLink.classList.add("mx-1");
-  deleteLink.addEventListener("click", (event) => {
-    event.preventDefault();
-    deleteClient(id);
-  });
-  const editLink = createEditLink(id);
-  div.appendChild(textPart1);
-  div.appendChild(mailLink);
-  div.appendChild(textPart2);
-  div.appendChild(editLink);
-  div.appendChild(deleteLink);
-
-  return div;
-}
-
-// <!-- trigger modal -->
-// <a href="" data-toggle="modal" data-target="#editClientModal"
-// 	>Edit client</a
-// >
-function createEditLink(id) {
-  const editLink = document.createElement("a");
-  editLink.innerHTML = "Edit";
-  editLink.setAttribute("href", "#");
-  editLink.setAttribute("data-toggle", "modal");
-  editLink.setAttribute("data-target", "#editClientModal");
-  editLink.setAttribute("data-client-id", id);
-  editLink.classList.add("mx-1");
-  editLink.classList.add("edit-client-link");
-  editLink.addEventListener("click", () => {
-    fillClientForm(id);
-  });
-
-  return editLink;
-}
-
-function fillClientForm(id) {
-  const currentClient = clients.find((client) => client.clientId == id);
-  if (editClientForm) {
-    editClientForm.firstName.value = currentClient.firstName;
-    editClientForm.lastName.value = currentClient.lastName;
-    editClientForm.email.value = currentClient.email;
-    editClientForm.gender.value = currentClient.gender;
-    editClientForm.amount.value = currentClient.amount;
-    editClientForm.date.value = currentClient.date;
-    editClientForm.clientID.value = id;
-    // editClientForm.photo.value = clients[id].avatar;
-    // "https://robohash.org/omnisveniamqui.jpg?size=50x50&set=set1";
-  }
-}
+const filterField = document.querySelector("#filterInput");
+filterField.addEventListener("keyup", (event) => {
+  filterList(event);
+});
 
 function editClient(form) {
   const data = {
@@ -159,29 +83,6 @@ function editClient(form) {
   // console.log("id-", id, data);
   if (id) updateDB(updates);
   hideModel("#editClientModal");
-}
-
-function deleteClient(id) {
-  // console.log("delete id -", id);
-  const currentClient = clients.find((client) => client.clientId == id);
-  console.log(
-    "Delete -" + currentClient.firstName + " " + currentClient.lastName
-  );
-
-  const captionClientNameDelete = document.querySelector(
-    "#captionClientNameDelete"
-  );
-  captionClientNameDelete.innerHTML =
-    currentClient.firstName + " " + currentClient.lastName;
-
-  $("#questDeleteClientModal").modal("show");
-  document
-    .querySelector("#questDeleteClientBtn")
-    .addEventListener("click", () => {
-      const clientRef = db.ref(`clients/${id}`);
-      clientRef.remove();
-      hideModel("#questDeleteClientModal");
-    });
 }
 
 function sortList(order) {
@@ -211,23 +112,8 @@ function sortList(order) {
 //   refreshData(sortedClients);
 // }
 
-function refreshData(updatedClients) {
-  clearList();
-  displayData(updatedClients);
-}
-
-function clearList() {
-  const ul = document.querySelector("#clientsData");
-  while (ul.firstChild) {
-    ul.removeChild(ul.firstChild);
-  }
-}
-
-function filterList() {
-  const filterString = document
-    .querySelector("#filterInput")
-    .value.toLowerCase()
-    .trim();
+function filterList(event) {
+  const filterString = event.target.value.toLowerCase().trim();
   if (filterString) {
     const filteredClients = clients.filter((client) => {
       return (
@@ -241,14 +127,23 @@ function filterList() {
     filteredClients.length === 0
       ? showResultListOrNotFound("showNotFound")
       : showResultListOrNotFound();
-    //   showNotFoundSection()
-    //   : showResultListSection();
   } else {
     refreshData(clients);
-    // showResultListSection();
     showResultListOrNotFound();
   }
 }
+
+const filterGender = [
+  { id: "filterFemale", value: "Female" },
+  { id: "filterMale", value: "Male" },
+];
+filterGender.forEach((gender) => {
+  const element = document.querySelector(`#${gender.id}`);
+  element.addEventListener("click", () => {
+    filterListGender(gender.value);
+  });
+});
+
 //--filterListGender  v1
 // function filterListGender(defaultGender) {
 //   const genderList = [];
@@ -264,59 +159,17 @@ function filterList() {
 
 //--filterListGender  v2
 function filterListGender(defaultGender) {
-  // const genderList = clients.filter(client => {
-  //   return client.gender.toLowerCase() == defaultGender.toLowerCase();
-  // });
   const genderList = clients.filter((client) => {
     return client.gender.toLowerCase() == defaultGender.toLowerCase();
   });
   refreshData(genderList);
 }
 
-function sumAmount(clientsList = clients) {
-  const total = clientsList.reduce((amount, client) => {
-    return amount + removeCurrencyFromAmount(client.amount);
-  }, 0);
-  document.querySelectorAll(".totalAmountContainer").forEach((element) => {
-    element.innerHTML = total.toFixed(2);
-  });
-}
-
 function removeCurrencyFromAmount(amount) {
   return amount ? Number(amount.slice(1)) : 0;
 }
 
-// function showNotFoundSection() {
-//   document.querySelector(".resultList").style.display = "none";
-//   document.querySelector(".notFound").style.display = "block";
-// }
-
-// function showResultListSection() {
-//   document.querySelector(".resultList").style.display = "block";
-//   document.querySelector(".notFound").style.display = "none";
-// }
-
-function showResultListOrNotFound(param = "showResultList") {
-  if (param == "showResultList") {
-    document.querySelector(".resultList").style.display = "block";
-    document.querySelector(".notFound").style.display = "none";
-  } else {
-    document.querySelector(".resultList").style.display = "none";
-    document.querySelector(".notFound").style.display = "block";
-  }
-}
-
 function addClient(form) {
-  // const data = {
-  //   firstName: "Demetris",
-  //   lastName: "Nerheny",
-  //   email: "dnerheny0@timesonline.co.uk",
-  //   gender: "Male",
-  //   amount: "$2.08",
-  //   date: "7/28/2019",
-  //   avatar: "https://robohash.org/omnisveniamqui.jpg?size=50x50&set=set1"
-  // };
-
   const data = {
     firstName: form.firstName.value,
     lastName: form.lastName.value,
@@ -328,30 +181,29 @@ function addClient(form) {
     // avatar: form.photo.value
   };
 
-  // console.log(data);
-
-  const newId = db.ref().child("clients").push().key;
-  // console.log(newId);
+  const newId = firebase.database().ref().child("clients").push().key;
 
   let updates = {};
   updates[`clients/${newId}`] = data;
 
   updateDB(updates);
-
   clearFieldsForm(form);
 }
 
 function updateDB(updates) {
-  db.ref().update(updates, function (error) {
-    if (error) {
-      console.error(
-        "New client was not added or was not saved! Error occured!"
-      );
-    } else {
-      // Data saved successfully!
-      console.log("Data added/saved to database!");
-    }
-  });
+  firebase
+    .database()
+    .ref()
+    .update(updates, function (error) {
+      if (error) {
+        console.error(
+          "New client was not added or was not saved! Error occured!"
+        );
+      } else {
+        // Data saved successfully!
+        console.log("Data added/saved to database!");
+      }
+    });
 }
 
 function clearFieldsForm(form) {
@@ -361,16 +213,70 @@ function clearFieldsForm(form) {
   form.amount.value = "";
 }
 
+const logOutBtn = document.querySelector("#logOut");
+logOutBtn.addEventListener("click", () => {
+  logOut();
+});
+
 function logOut() {
   firebase
     .auth()
     .signOut()
     .then(() => {
       // Sign-out successful.
-      window.location.href = "login.html";
+      window.location.href = "./login.html";
     })
     .catch((error) => {
       // An error happened.
       console.error(error);
+    });
+}
+
+//-----------------export--------------------------
+
+export function sumAmount(clientsList = clients) {
+  const total = clientsList.reduce((amount, client) => {
+    return amount + removeCurrencyFromAmount(client.amount);
+  }, 0);
+  document.querySelectorAll(".totalAmountContainer").forEach((element) => {
+    element.innerHTML = total.toFixed(2);
+  });
+}
+
+export function fillClientForm(id) {
+  const currentClient = clients.find((client) => client.clientId == id);
+  if (editClientForm) {
+    editClientForm.firstName.value = currentClient.firstName;
+    editClientForm.lastName.value = currentClient.lastName;
+    editClientForm.email.value = currentClient.email;
+    editClientForm.gender.value = currentClient.gender;
+    editClientForm.amount.value = currentClient.amount;
+    editClientForm.date.value = currentClient.date;
+    editClientForm.clientID.value = id;
+    // editClientForm.photo.value = clients[id].avatar;
+    // "https://robohash.org/omnisveniamqui.jpg?size=50x50&set=set1";
+  }
+}
+
+export function deleteClient(id) {
+  // console.log("delete id -", id);
+  const currentClient = clients.find((client) => client.clientId == id);
+  console.log(
+    "Delete -" + currentClient.firstName + " " + currentClient.lastName
+  );
+
+  const captionClientNameDelete = document.querySelector(
+    "#captionClientNameDelete"
+  );
+  captionClientNameDelete.innerHTML =
+    currentClient.firstName + " " + currentClient.lastName;
+
+  $("#questDeleteClientModal").modal("show");
+  document
+    .querySelector("#questDeleteClientBtn")
+    .addEventListener("click", () => {
+      const clientRef = firebase.database().ref(`clients/${id}`);
+      clientRef.remove();
+      hideModel("#questDeleteClientModal");
     });
 }
